Added coverage
